+++ src/pagesComponents/createNewPassword/ui/CreateNewPassword.tsx
@@ -7,8 +7,8 @@ import { useTranslation } from 'react-i18next';
 import { CreateNewPasswordFormValues, createNewPasswordSchema } from '@/features/schemas/authSchema/authShema';
 import { Button } from '@/shared/button/Button';
 import { Card } from '@/shared/card/Card';
-import { FormTextfield } from '@/shared/forms/FormTextfield';
 import { Typography } from '@/shared/ui';
+import { FormTextfield } from '@/shared/ui/forms/FormTextfield';
 import { zodResolver } from '@hookform/resolvers/zod';
 
 import s from '@/pagesComponents/createNewPassword/ui/CreateNewPassword.module.scss';
