+++ src/pagesComponents/signIn/ui/SignIn.tsx
@@ -4,8 +4,8 @@ import React from 'react';
 import useSignInForm from '@/pagesComponents/signIn/lib/hooks/useSignInForm';
 import { Button } from '@/shared/button/Button';
 import { Card } from '@/shared/card/Card';
-import { FormTextfield } from '@/shared/forms/FormTextfield';
 import { Typography } from '@/shared/ui';
+import { FormTextfield } from '@/shared/ui/forms/FormTextfield';
 import HeadSignInAndSignUp from '@/widget/headSignInAndSignUp/HeadSignInAndSignUp';
 import Link from 'next/link';
