+++ src/shared/ui/input/Input.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Input } from '@/shared/ui/input/Input';
+import { Input } from '@/shared/ui';
 
 const meta = {
   component: Input,
