+++ src/shared/ui/modal/Modal.stories.tsx
@@ -2,8 +2,7 @@ import { useState } from 'react';
 
 import ImageOutline from '@/../public/assets/components/ImageOutline';
 import { Button } from '@/shared/button/Button';
-import { Input } from '@/shared/ui/input/Input';
-import { Modal, Props as ModalProps } from '@/shared/ui/modal/Modal';
+import { Input, Modal, Props as ModalProps } from '@/shared/ui';
 import { Checkbox } from '@radix-ui/react-checkbox';
 import { Meta, StoryFn } from '@storybook/react';
