+++ src/shared/ui/pagination/ui/PaginationWithSelect.stories.tsx
@@ -4,7 +4,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import { useState } from 'react';
 
-import { PaginationWithSelect, selectOptionsType } from '@/shared/ui/pagination/ui/PaginationWithSelect';
+import { PaginationWithSelect, selectOptionsType } from '@/shared/ui';
 
 const meta = {
   argTypes: {},
