+++ src/shared/ui/textArea/TextArea.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import TextArea from '@/shared/ui/textArea/TextArea';
+import { TextArea } from '@/shared/ui';
 
 const meta = {
   component: TextArea,
