+++ src/shared/ui/typography/Typography.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Typography } from '@/shared/ui/typography/Typography';
+import { Typography } from '@/shared/ui';
 
 const meta = {
   argTypes: {
